+++ src/senaite/core/interfaces/catalog.py
@@ -64,3 +64,13 @@ class IWorksheetCatalog(ISenaiteCatalogObject):
 class IReportCatalog(ISenaiteCatalogObject):
     """Marker interface for Senaite report catalog
     """
+
+
+class ILabelCatalog(ISenaiteCatalogObject):
+    """Marker interface for Senaite label catalog
+    """
+
+
+class IClientCatalog(ISenaiteCatalogObject):
+    """Marker interface for Senaite client catalog
+    """